Change gamesform to presentational component
